% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ExperimentManager.R
\name{initExperiment}
\alias{initExperiment}
\title{Initialize Experiment}
\usage{
initExperiment(expPath)
}
\arguments{
\item{expPath:}{the experiment absolute path}
}
\description{
This function initialize the experiment. Creates a folder structure at the
specified path defined by <expPath>
}
\examples{
initExperiment()
}
\keyword{ExperimentManager}

