% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ExperimentManager.R
\name{runExp}
\alias{runExp}
\title{Run Experiment}
\usage{
runExp(d, file, description)
}
\arguments{
\item{d:}{the experiment definition; file the script filename; description: a string description of the run features}
}
\description{
This function creates a new experiment run:
 1. makes a backup of the script code
 2. stores results
 3. stores results and additional information
}
\examples{
runExp()
}
\keyword{ExperimentManager}

